% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_names.R
\docType{data}
\name{example_names}
\alias{example_names}
\title{Example names}
\format{
A vector of strings
}
\source{
Star Wars
}
\usage{
example_names
}
\description{
An example data set containing six names from the Star Wars universe
}
\keyword{datasets}
