% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makegroups.R
\name{make_groups}
\alias{make_groups}
\title{Make groups of 2 persons}
\usage{
make_groups(names)
}
\arguments{
\item{names}{The vector of names}
}
\value{
The re-arranged matrix of names
}
\description{
Randomly arranges a vector of names into a matrix with 2 columns and
whatever number of columns is required
}
